Corrige o CI: flags como constantes e a porta do teste como uint16_t
O clang-tidy recusa um enum de bitmask (performance-enum-size), e um
conjunto de flags nao e mesmo uma escolha de um valor: viram constantes
inline. E o MSVC do CI trata como erro o 1 (int) que make_unique passava
para a porta uint16_t do cliente de teste sem servidor.

+++ tests/integration/test_call_session.cpp
@@ -7,6 +7,7 @@
 
 #include <atomic>
 #include <chrono>
+#include <cstdint>
 #include <memory>
 #include <mutex>
 #include <string>
@@ -572,7 +573,10 @@ class CallSessionTest : public ::testing::Test {
   /// A client pointed at a port nothing listens on: what the login screen
   /// looks like when the server is down.
   Client& add_without_a_server(const std::string& username) {
-    clients_.push_back(std::make_unique<Client>(1, username));
+    // Typed as the port it is: make_unique forwards a plain 1 as an int, and
+    // the narrowing to the constructor's uint16_t is an error on MSVC with
+    // warnings as errors.
+    clients_.push_back(std::make_unique<Client>(static_cast<std::uint16_t>(1), username));
     return *clients_.back();
   }
